[DM-24930] Move XForwardedFiltered to its own file
Hopefully this won't be needed in the long run if upstream takes
this code.  Even if it is, it's different than the application
setup logic, so it's clearer to keep it in a separate file.
