+++ src/gafaelfawr/app.py
@@ -13,8 +13,6 @@ import aioredis
 import jinja2
 from aiohttp import web
 from aiohttp.web import Application
-from aiohttp_remotes.exceptions import RemoteError
-from aiohttp_remotes.x_forwarded import XForwardedBase
 from aiohttp_session.cookie_storage import EncryptedCookieStorage
 from cachetools import TTLCache
 from safir.http import init_http_session
@@ -25,13 +23,13 @@ from structlog import get_logger
 
 from gafaelfawr.config import Config
 from gafaelfawr.handlers import init_routes
+from gafaelfawr.x_forwarded import XForwardedFiltered
 
 if TYPE_CHECKING:
     from aiohttp import ClientSession
     from aioredis import Redis
-    from ipaddress import _BaseNetwork
     from structlog import BoundLogger
-    from typing import Any, Awaitable, Callable, Dict, Optional, Sequence
+    from typing import Any, Awaitable, Callable, Optional
 
     Handler = Callable[[web.Request], Awaitable[web.StreamResponse]]
 
@@ -97,89 +95,6 @@ async def create_app(
 
     logger.info("Starting")
     return app
-
-
-class XForwardedFiltered(XForwardedBase):
-    """Middleware to update the request based on ``X-Forwarded-For``.
-
-    The semantics we want aren't supported by either of the
-    :py:mod:`aiohttp_remotes` middleware classes, so we implement our own.
-    This is similar to `~aiohttp_remotes.XForwardedRelaxed` except that it
-    takes the rightmost IP address that is not contained within one of the
-    trusted networks.
-
-    Parameters
-    ----------
-    trusted : Sequence[Union[`ipaddress.IPv4Network`, `ipaddress.IPv6Network`]]
-        List of trusted networks that should be skipped over when finding the
-        actual client IP address.
-    """
-
-    def __init__(self, trusted: Sequence[_BaseNetwork]):
-        self._trusted = trusted
-
-    @web.middleware
-    async def middleware(
-        self, request: web.Request, handler: Handler
-    ) -> web.StreamResponse:
-        """Replace request information with details from proxy.
-
-        Honor ``X-Forwarded-For`` and related headers.
-
-        Parameters
-        ----------
-        request
-            The aiohttp.web request.
-        handler
-            The application's request handler.
-
-        Returns
-        -------
-        response
-            The response with a new ``logger`` key attached to it.
-
-        Notes
-        -----
-        The remote IP address will be replaced with the right-most IP address
-        in ``X-Forwarded-For`` that is not contained within one of the trusted
-        networks.  The last entry of ``X-Forwarded-Proto`` and the contents of
-        ``X-Forwarded-Host`` will be used unconditionally if they are present
-        and ``X-Forwarded-For`` is also present.
-        """
-        try:
-            # https://github.com/python/mypy/issues/8772
-            overrides: Dict[str, Any] = {}
-            headers = request.headers
-
-            forwarded_for = list(reversed(self.get_forwarded_for(headers)))
-            if not forwarded_for:
-                return await handler(request)
-
-            for ip in forwarded_for:
-                if any((ip in network for network in self._trusted)):
-                    continue
-                overrides["remote"] = str(ip)
-                break
-
-            # If all the IP addresses are from trusted networks, take the
-            # left-most.
-            if "remote" not in overrides:
-                overrides["remote"] = str(forwarded_for[-1])
-
-            proto = self.get_forwarded_proto(headers)
-            if proto:
-                overrides["scheme"] = proto[-1]
-
-            host = self.get_forwarded_host(headers)
-            if host is not None:
-                overrides["host"] = host
-
-            request = request.clone(**overrides)
-            return await handler(request)
-
-        except RemoteError as exc:
-            exc.log(request)
-            return await self.raise_error(request)
 
 
 async def setup_middleware(app: Application, config: Config) -> None:
